requirements.txt
feedparser>=6.0.0
requests>=2.31.0
python-dateutil>=2.8.0
lxml>=4.9.0
beautifulsoup4>=4.12.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
hypothesis>=6.0.0
